pyproject.toml
[project]
name = "kartopu-blog"
version = "0.1.0"
description = "Financial blog project about FIRE and personal finance."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django~=6.0.0",
    "django-imagekit>=6.0.0",
    "django-types>=0.22.0",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.2.13",
]

[dependency-groups]
dev = [
    "django-upgrade>=1.29.1",
]
